Added Button (WIP)

+++ Engine/Source/Random.h
@@ -19,7 +19,7 @@ inline Vector2 randomOnUnitCircle()
 	return Vector2{ Math::Cos(Math::DegToRad(angle)), Math::Sin(Math::DegToRad(angle)) };
 }
 
-inline Vector2 randomOnUnitCircle(int startDegree, int endDegree)
+inline Vector2 randomOnUnitCircle(float startDegree, float endDegree)
 {
 	float angle = randomf(startDegree, endDegree);
